Media_director/requirements.txt
flask
python-telegram-bot
uwsgi
requests